requirements.txt
fastapi[all]
Pillow
numpy
SimpleITK
pandas
openpyxl
pathlib
python-dotenv
pynrrd